Initializes React Native app
Sets up a basic React Native application using Expo.

- Creates essential files such as `app.json` and `App.tsx`.
- Adds necessary dependencies to `package.json`.
- Configures `.gitignore` to exclude generated native folders.
- Removes obsolete configuration files and example code.
