--- apps/mobile/.vscode/extensions.json
@@ -1,1 +1,0 @@
-{ "recommendations": ["expo.vscode-expo-tools"] }
