+++ apps/mobile/assets/adaptive-icon.png
